etape7-mlops/requirements.txt
# MLOps Dependencies pour Vertex AI Pipelines
google-cloud-aiplatform==1.38.0
kfp==2.4.0

# Data processing
pandas==2.1.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.2

# Deep Learning (pour BERT)
transformers==4.35.2
torch==2.1.1

# Cloud Storage
google-cloud-storage==2.10.0

# Utilities
joblib==1.3.2
tqdm==4.66.1
